+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # R0.3 Comming soon
 
 - Add preview for json files
+- Migration from Bower to Yarn
 - Maj Electron version 1.7.9 -> 1.7.10
 - Fix : Problem with count translation return 0 when the files are not empty